blockchain/blockchain_indexer/requirements.txt
# Blockchain Indexer Requirements

# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.9

# Retry logic
tenacity>=8.2.3

# Logging (optional, for enhanced logging)
python-json-logger>=2.0.7

# Development dependencies (optional)
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0
